% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{auc}
\alias{auc}
\title{auc
auc runs the whywhere Area Under Curve function}
\usage{
auc(p0, pa)
}
\arguments{
\item{p0}{vector of probabilities}

\item{pa}{vector of 1's and 0's for presence and absence}
}
\value{
value AUC
}
\description{
auc
auc runs the whywhere Area Under Curve function
}

