% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{dseg}
\alias{dseg}
\title{dseg}
\usage{
dseg(file, ext, Basis, pa)
}
\arguments{
\item{ext}{a geographic extent}

\item{Basis}{Presence locations and absence locations}

\item{pa}{a vector of zeros and ones for presence and absence locations}

\item{files}{A single environmental file}
}
\value{
A model object membership
}
\description{
dseg generates the ww model - a segmented distribution
}

