% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{getraster}
\alias{getraster}
\title{getraster
getraste gets a raster using raster package, and added
global range for pgm files}
\usage{
getraster(file)
}
\arguments{
\item{file}{file name}
}
\value{
raster
}
\description{
getraster
getraste gets a raster using raster package, and added
global range for pgm files
}

