% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{membership}
\alias{membership}
\title{membership}
\usage{
membership(file, ext, Basis, pa)
}
\arguments{
\item{ext}{a geographic extent}

\item{Basis}{Presence locations and absence locations}

\item{pa}{a vector of zeros and ones for presence and absence locations}

\item{files}{A single environmental file}
}
\value{
A model object membership
}
\description{
membership is used to develop an environmental model from known locations,
and use them to predict further occurrences
}
\details{
The main application is species niche modelling.
The algorithm implemented here claims to be both more accurate and
more informative than other methods (hence the name 'WhyWhere').
}

