% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{model.ww}
\alias{model.ww}
\title{model.ww
model.ww develops the model}
\usage{
model.ww(Pres, files, multi = F, limit = 0, beam = 5, e = 0.5,
  plot = FALSE)
}
\arguments{
\item{Pres}{A 2 column data.frame of known locations}

\item{files}{The environmental files for raster}

\item{multi}{multiple dimensions}

\item{limit}{minimum for entry into results}

\item{beam}{number of entries to keep in table}

\item{e}{size of border around presence points}

\item{plot}{logical}
}
\value{
scaled vector
}
\description{
model.ww
model.ww develops the model
}

