% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{range01}
\alias{range01}
\title{range01
range01 adjusts range to equal area}
\usage{
range01(x)
}
\arguments{
\item{x}{vector}
}
\value{
scaled vector
}
\description{
range01
range01 adjusts range to equal area
}

