% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{whywhere}
\alias{whywhere}
\title{whywhere}
\usage{
whywhere(Pres, files, multi = F, limit = 0, beam = 5, e = 0.5,
  plot = FALSE)
}
\arguments{
\item{Pres}{A 2 column data.frame of known locations}

\item{files}{The environmental files for raster}

\item{multi}{multiple dimensions}

\item{limit}{minimum for entry into results}

\item{beam}{number of entries to keep in table}

\item{e}{size of border around presence points}

\item{plot}{logical}
}
\value{
result A frame of results
}
\description{
whywhere is used to develop an environmental model from known locations,
and use them to predict further occurrences
}
\details{
The main application is species niche modelling.
The algorithm implemented here claims to be both more accurate and
more informative than other methods (hence the name 'WhyWhere').
}

