% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww2.R
\name{ww}
\alias{ww}
\title{ww}
\usage{
ww(Pres, files)
}
\arguments{
\item{Pres}{A 2 column data.frame of known locations}

\item{files}{The environmental files for raster}
}
\value{
result A frame of results
}
\description{
ww is used to develop an environmental model from known locations,
and use them to predict further occurrences.
The main application is species niche modelling.
}
\examples{
result=ww(Bradypus_Pres,Bradypus_files)
print(result)
}

