% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww1.R
\name{membership}
\alias{membership}
\title{membership}
\usage{
membership(file, P, A, factor = F, extent)
}
\arguments{
\item{factor}{Is a factor variable}

\item{extent}{The extent of the range of values}

\item{files}{An environmental file}

\item{Pres}{An array of presence coords}

\item{Back}{Absence or background coords}
}
\value{
A membership function for that variable
}
\description{
membership is used to develop an environmental model from known locations,
and use them to predict further occurrences
}
\details{
The main application is species niche modelling.
The algorithm implemented here claims to be both more accurate and
more informative than other methods (hence the name 'WhyWhere').
}

