% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ww1.R
\name{predict.membership}
\alias{predict.membership}
\title{predict}
\usage{
predict.membership(files, models, extent)
}
\arguments{
\item{pa}{A vector of presence/absence}

\item{data}{A matrix of memberships for each env value}

\item{dimensions}{number of terms}
}
\value{
Possibilities arranged in order of probability
}
\description{
predict is used to develop an environmental model from known locations,
and use them to predict further occurrences
}
\details{
The main application is species niche modelling.
The algorithm implemented here claims to be both more accurate and
more informative than other methods (hence the name 'WhyWhere').
}

